[MOV] lunch: Split model files according to guidelines
