+++ addons/lunch/models/__init__.py
@@ -6,6 +6,8 @@ from . import lunch_cashmove
 from . import lunch_location
 from . import lunch_order
 from . import lunch_product
+from . import lunch_product_category
+from . import lunch_topping
 from . import lunch_supplier
 from . import res_company
 from . import res_config_settings
